Pushed normalisation of scaling units into term_hspan().

+++ term.h
@@ -103,7 +103,7 @@ struct	termp {
 	void		(*endline)(struct termp *);
 	void		(*advance)(struct termp *, size_t);
 	size_t		(*width)(const struct termp *, char);
-	size_t		(*hspan)(const struct termp *,
+	double		(*hspan)(const struct termp *,
 				const struct roffsu *);
 	const void	 *argf;		/* arg for headf/footf */
 	union {
